Moves files to new package
