+++ tests/com/tkomiya/tests/TranslationJsonConverterTest.java
@@ -9,9 +9,9 @@ import org.junit.Test;
 
 import com.tkomiya.infrastructure.TranslationJsonConverter;
 import com.tkomiya.testutils.TestTranslationLoader;
-import com.tkomiya.testutils.datageneration.JapaneseTranslationFactory;
-import com.tkomiya.testutils.datageneration.KoreanTranslationFactory;
-import com.tkomiya.testutils.datageneration.EnglishTranslationFactory;
+import com.tkomiya.testutils.datageneration.translation.EnglishTranslationFactory;
+import com.tkomiya.testutils.datageneration.translation.JapaneseTranslationFactory;
+import com.tkomiya.testutils.datageneration.translation.KoreanTranslationFactory;
 
 public class TranslationJsonConverterTest {
